# shopify/api/build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"RedirectUrl": {
				"properties": {
					"redirect_url": {
						"type": "string"
					}
				},
				"required": [
					"redirect_url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InitiatePaymentDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"gid": {
						"type": "string"
					},
					"group": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string"
					},
					"test": {
						"type": "boolean"
					},
					"merchant_locale": {
						"type": "string"
					},
					"payment_method": {},
					"cancel_url": {
						"type": "string"
					},
					"proposed_at": {
						"type": "string"
					},
					"customer": {},
					"kind": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"gid",
					"group",
					"amount",
					"currency",
					"test",
					"merchant_locale",
					"payment_method",
					"cancel_url",
					"proposed_at",
					"customer",
					"kind"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			}
		}
	},
	"info": {
		"title": "@solana/pay-shopify",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/onboarding/install": {
			"get": {
				"operationId": "Install",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "shop",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/onboarding/redirect": {
			"get": {
				"operationId": "Redirect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "shop",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/onboarding": {
			"get": {
				"operationId": "GetSession",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "onboardSessionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/onboarding/wallet": {
			"post": {
				"operationId": "SaveWallet",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"api_key": [
							"onboarding"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/payment-session": {
			"get": {
				"operationId": "GetPaymentSession",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "paymentSessionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/payment-session/initiate": {
			"post": {
				"operationId": "InitiatePayment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RedirectUrl"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InitiatePaymentDto"
							}
						}
					}
				}
			}
		},
		"/payment-session/refund": {
			"post": {
				"operationId": "Refund",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/security/jwks.json": {
			"get": {
				"operationId": "GetJwks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"keys": {
											"items": {},
											"type": "array"
										}
									},
									"required": [
										"keys"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/test/initiate": {
			"get": {
				"operationId": "Initiate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/test/onboard": {
			"get": {
				"operationId": "Onboard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}